move log_access out of own module and into base commands module
